requirements.txt
# PickyText — requirements.txt
# Python 3.13 | Windows 11 only

# GUI
PyQt6>=6.7.0

# Screenshot
mss>=9.0.1

# Global hotkey
keyboard>=0.13.5

# OCR — Windows WinRT bindings
winsdk==1.0.0b10

# Image processing
Pillow>=10.4.0
numpy>=2.0.0

# Translation HTTP client
httpx>=0.27.0

# Auto-update version comparison
packaging>=24.0

# ── Optional components (install manually if needed) ──────────────────────────
# Tesseract OCR fallback:
#   pip install pytesseract>=0.3.13
#   + install Tesseract binary from https://github.com/UB-Mannheim/tesseract/wiki

# Argos Translate (local offline translation fallback):
#   pip install argostranslate>=1.9.6